% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimarathon.R
\name{get_results}
\alias{get_results}
\title{Get Teamstaffel Results Table}
\usage{
get_results(
  sex = "W",
  ak_class = "G",
  year = 2022,
  base_url = "https://berlin.r.mikatiming.com"
)
}
\arguments{
\item{sex}{either "M" for men or "W" for women default: "W"}

\item{ak_class}{"G" (Grundschule), "O" (Oberschule), "SO" (Sonderwertung),
default: "G"}

\item{year}{default: 2022}

\item{base_url}{default: "https://berlin.r.mikatiming.com"}
}
\value{
tibble with results
}
\description{
Get Teamstaffel Results Table
}
\examples{
women <- get_results("W")
women
}
